JPA integration and files corrections
