+++ src/main/java/aplication/Program.java
@@ -2,6 +2,9 @@ package aplication;
 
 
 
+//import entities.AccountEntity;
+
+
 
 /*
  * Program is responsible to run the application 
@@ -10,18 +13,19 @@ package aplication;
 public class Program {
 
 	
-	
-	public static void main(String[] args) {
+	public static void main(String[] args) throws Exception {
 		
-					
-		//start controller to save into txt
-		//Controller controller = new Controller();
+		/*			
+		start controller to save into txt
+		Controller controller = new Controller();
 		
-		//start controller to save into mysql
+		start controller to save into mysql using jdbc
 		ControllerUsingMySQL controller = new ControllerUsingMySQL();
-			
-			System.out.flush();
+		*/
+		
 		
 	}
+	
+	
 
 }
