+++ src/main/java/database/JDBC/DAOAccount.java
@@ -1,4 +1,4 @@
-package database_JDBC;
+package database.JDBC;
 
 
 import java.sql.PreparedStatement;
