+++ src/main/java/database/JPA/UserDAO.java
@@ -1,4 +1,4 @@
-package database_JPA;
+package database.JPA;
 
 import java.util.List;
 
@@ -8,20 +8,24 @@ import javax.transaction.Transactional;
 
 import entities.User;
 
-public class DAOUser extends DAO{
+public class UserDAO extends DAO{
 	@Transactional
-	/**
-	 * Find an user by ID
-	 * @param id
-	 * @return
-	 */
+	
 	public User findbyId(Integer id) {
 		User user = entityManager.find(User.class, id);
 		return user;
 	}
+	public User findByCPF(String CPF) {
+		String jpql = "select u from User u where u.CPF = : userCPF";
+		TypedQuery<User> typedQuery = entityManager.createQuery(jpql, User.class);
+		typedQuery.setParameter("userCPF", CPF);
+		User user = typedQuery.getSingleResult();
+		return user;
+	}
 
 	public List<User> allUsers(EntityManager entityManager) {
-		String jpql= "select u from User u";
+		String jpql= "select new entities.User(fullname,CPF,birthday,id)"
+				+ " from User";
 		TypedQuery<User> typedQuery = entityManager.createQuery(jpql, User.class);
 		List<User> list = typedQuery.getResultList();
